% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiased_sparse_pca.R
\name{alternating_debiased_sparse_pca}
\alias{alternating_debiased_sparse_pca}
\title{De-biased Sparse PCA}
\usage{
alternating_debiased_sparse_pca(
  X,
  beta_hat,
  lambda,
  T,
  max_iter = 100,
  tol = 1e-06
)
}
\arguments{
\item{X}{Numeric matrix of size \(n \times p\).}

\item{beta_hat}{Initial guess for the principal component vector.}

\item{lambda}{Regularization parameter.}

\item{T}{Constraint on \(L_1\)-norm.}

\item{max_iter}{Maximum number of iterations (default = 100).}

\item{tol}{Convergence threshold (default = 1e-06).}
}
\value{
A list with \code{beta_debiased}, \code{eigenvalue}, and \code{Theta}.
}
\description{
Implements the de-biased sparse PCA algorithm.
}
